chore: simplify and rename
